+++ rascaline/src/labels/mod.rs
@@ -2,9 +2,7 @@ mod samples;
 
 pub use self::samples::{SpeciesFilter, SamplesBuilder};
 pub use self::samples::AtomCenteredSamples;
-pub use self::samples::SamplesPerAtom;
 pub use self::samples::LongRangeSamplesPerAtom;
-pub use self::samples::Structures;
 
 mod keys;
 pub use self::keys::KeysBuilder;
